requirements.txt
aiogram==3.13.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.6.1
gspread-asyncio==1.10.0
google-auth==2.37.0
alembic==1.14.0
httpx==0.27.2
python-dotenv==1.0.1
